Add flush to NVDEC decoding process
When the input image has an issue, like outside of the supported
resolution, then the parser of NVDEC decoder is initialized but,
the `handle_video_sequence` callback, which is supposed to be called
once par new input, is not called.

It turned out that if two images with unsupported resolutions are
fed, then the NvDec decoder assumes these frames are coming from
the same video, and it does not call `handle_video_sequence` callback
for the second time, even though decoder is not initialized on our end.

This commit fixes the issue by introducing `flush` method and calling
it everytime before the decoding starts.

TODO: Measure the impact of calling `flush` every time. If it is not
trivial, then call it only when the previous decoding fails.