Refactor BSF for NVDEC demuxing
NVDEC seems to require H264/H265 to be in annex B form.
This is handled by applying `toannexb` bit stream filter.

Originally, to reduce the memory consumption, this BSF was
inserted to the demux function, and the Packets returned by
demux function only contains the filtered packets.
However, this required to have two separate demuxing function,
which are almost identical except BSF.
This is not great for composability of async demuxing/decoding.
Therefore, we make the BSF application separate step as demuxing,
this way, the new async composition can use the BSF dynamically,
depending on if the input data is supported by NVDEC.

The downside is the increased CPU memory consumption. Decoding time
should not be a concern as long as the BSF application happens
on demuxer thread.