Change naming scheme for async functions (#26)
1. Use async prefix so that async functions are grouped in alphabetical
order. (async functions and synchronous functions are not 1to1 mapping
because async functions have helper async functions that are not public
API for sync functions. For example, `async_apply_bsf`)

2. Functions that deal with similar type with different template
arguments are registered with the same name, so that Python choose the
right function at runtime, while reducing the number of public APIs.