Detach Packets/Frames lifetimes from those of their Python vars (#28)
The previous way of implementing async process had one problem. The
lifetime of Packets/Frames are tied to the Python variables, which means
that Python variables would not give up the ownership of the memory, and
those memories are deallocated when and only when the variables are
garbage-collected by Python.

This means that memory deallocations only happen on a Python thread, and
it blocks the async loop while memory deallocation happens.