Do not push nullptr to Packets
nullptr is used to let Audio/Video decoder know that the stream
is ended.

This is not required for NVDEC or Image decoder.
Originally, the demuxer was developed for FFmpeg, so it automatically
added nullptr.
Now that there are other decoders, we push the nullptr to packet
vector only when it's necessary, that is right before the packet
vector is passed to audio/video decoder.