Reduce the log level for cancellation
When top-level async task is cancelled, multiple of tasks are
cancelled simulutaneously and the cancellations propagate to
C++ code.

Currently, cancellations emit logs like

```
future.h:39] coroutine operation cancelled
```

and it is not so useful when so many of the tasks are cancelled at
the same time.

So we use different log level for cancellataion.