Support zero-copy bytes demuxing (#40)
PyBind11 can pass `bytes` to C++ without creating a copy when the C++
receives it as `std::string_view`.
(`str` object cannot be passed to C++ without creating a copy, and
PyBind11 creates a temporary copy, which is deleted after the function
call, which is rather annoying when writing async functions.)

This commit makes demuxing functions to switch the implementation based
on the type of src.
If the src is `str`, it calls the function with copy. If the src is
`bytes`, it calls the function without copy.

The intended use for `bytes` is that `bytes` is in-memory media data,
rather than an URI, so it will also use `BytesAdoptor` automatically.

On C++ side, the passed arguments are copied regardless of `std::string`
or `std::string_view` when it go through `folly::coro::co_invoke`. It's
only when the argument type is `string_view` that the caller must keep
the object the view is pointing alive.