Replace PyBind11 with Nanobind (#44)
Nanobind has a better semantics for data ownerships and trasnfer of it.
Passing `unique_ptr` automatically transfers the onwership of the data
and makes the Python variable invalid state.

This is what we implemented with Wrapper classes, thus the adoption of
Nanobind allows to get rid of these wrapper classes from libspdl.

Regarding the speed, the following is the results of running ImageNet
validataion. I ran them in turn.

Before (pybind11 2.10.4)
- FPS=3473.62 (49920/14.37)
- FPS=3527.55 (49920/14.15)
- FPS=3549.92 (49920/14.06)
- FPS=3661.70 (49920/13.63)

After (nanobind 1.5.2)
- FPS=3426.72 (49920/14.57)
- FPS=3378.06 (49920/14.78)
- FPS=3496.04 (49920/14.28)
- FPS=3562.37 (49920/14.01)

It seems nanobind is slightly slower than pybind? But they are in
acceptable range, so we will go with nanobind.

Other notable changes are

1. Passing memoryview is removed.