Deprecate functions take media_type as arg
Functions like `demux_media(media_type)` are used so that the
maintenance effort due to the number of API is small.

With the upcoming introduction of GIL-free binding, we can get
rid of a lot of boilerplate code, so the maintenance effort will
be lower.

It is better to have these functions separated based on media type.