Good bye Folly (#67)
This project started with folly.
In fact the repository was called folly-subset, aiming to build folly
with minimum dependency. So the first commits were all about building
folly with custom CMake file, and it took like a week to make it work.

I picked folly so that I can experiment with coroutine and thread pool
for high performance data loading. Folly served that purpose very well
and we learned that it is possible to do high-throughput data loading.

Now the project has evolved and we learned a bunch of new techniques,
and we got to the point where we can use thread pool in Python.

Also we managed to roll our own coroutine-based generator to FFmpeg's
decoding mechanism, which is triple-nested while loop.

Folly is really a great library, but its use came to the end for SPDL.
As we are paving the way for open-source the project, it is better to
have less dependencies.

So this commit removes Folly from SPDL.

Thank you Folly!