Support regular generator in pipe
This commit adds support for generator function to pipe.

The main intended usage is to run generator in subprocess, and
get the result one-by-one.

However, since generator objects are not picklable, it is not
possible to run each `next(agen)` invocation in process,
therefore when generator is passed, it is converted to a regular
function, so that the results are returned only after the generator
is exhausted.

It should be technically possible to remove this restriction when
the generator is executed in threads instead of process, but I leave
it to the follow-up.