Add function to load NPY without copy
When loading numpy array from bytes, `np.load(BytesIO(data))` is
typically used.
The `np.load` function is implemented in a way it supports
file-like object well. It reads data chunk-by-chunk.

When we have the array data as bytes, this is inefficient.
We can create an array without copy by parsing the header
(shape and dtype), then interpretating the data region.