Add new (experimental) build configs (#382)
We are splitting the packaging into core (pure Python) and io (with C++
and possibly with CUDA).
Due to the nature of monorepo, it is challenging to have multiple
packaging config that shares the same source code root.
Turned out that you can use symlink when building a wheel.
The `uv` package manager can handle multiple packages with inter
dependencies.

So we develop the new way to build and distribute in a different
location than the current one.

The new way to build packages are.

```
uv build --wheel --all-packages --directory packaging
```

or

```
./packaging/build_wheels.sh
```