Make SKIP internal to pipe implementation
Rather than pass the SKIP sentinel to the next pipe, it is better
to not to pass it at all.

SKIP logic is currently used only by Aggregate, so for the sake
of simplifying the implementation, make the SKIP internal to pipe,
and make source and sink unaware of it.