Fix buffer overrun (#466)
When FFmpeg requests to seek on the byte string we pass, we did not
consider the case where the current position + offset can be outside of
the memory region.

Also we did not consider the possibility that when this happens, (buffer
size - position) can be negative due to this, and passing negative value
to `memcpy`, which can cause SEGV.

Also fixes typo.