Use ffmpeg 7.0.2 for build (#568)
FFmpeg 7.1 deprecated `AVCodec::pix_fmts` in lieu of newly introduced
`avcodec_get_supported_config`.
The build process is issuing bunch of deprecation warnings, and we want
to get rid of it, but if we migrate to `avcodec_get_supported_config`,
then the binary won't work with FFmpeg 7.0.

Therefore we use FFmpeg 7.0 in our build process.