Increase the pipe output buffer size to 2 (#700)
Currently the size of pipe output buffer is 1. This means that even when
the upstream is fast enough, unless the queue becomes empty once, the
next item cannot be placed.

This makes it impossible for the queue data readiness to reach 100%. So
we increase it to 2 so that queue can be always non-empty as long as the
upstream is fast enough.