Propagate severe Pipeline errors to front-end (#699)
This commit changes the behavior when Pipeline encounters an issue in
the background.
When an exception that affects the execution of the Pipeline occurs,
such exception is stored and later propagated back to the front end, so
that front end can notice and react to such an error.

The error includes
- Whatever failure source encounters.
- Pipe function signature issue.
- Task failures occurred more than the specified threshold.

The regular task failures are absorbed thus not propagated back.

closes #268