Revise the way periodic task is terminated
Originally, the periodic report task was simply cancelled.

However, that is not enough. One has to await the task
after the cancellation, then handle the outcome.

https://discuss.python.org/t/asyncio-cancel-a-cancellation-utility-as-a-coroutine-this-time-with-feeling/26304

I am not sure if cancel and await is enough as we used
`while True` loop in the periodic task.
Also, having a complex error handling logic in the `finally`
clause feels not simple enough.

So we change the implementation of _periodic_dispatch, so that
the signal is stop is sent via Event.
Also get rid of the fire-and-forget idiom.