Update shuffle (#857)
Summary:

When `IterableWithShuffle` is converted to `Iterable` with `embed_shuffle`, the expected behavior is each iteration generates the different sequence. So that they behave more like a random generator.

For this the seed value needs to be updated after each shuffling.
The current implementation delays the increment until the end of iteration.

Differential Revision: D80646907