Fix the handling of StopAsyncIterator for FailCounter
Summary:
When `max_failure` is enabled, a hook which counts the number of exception raised is installed.

This hook should not count `StopAsyncIterator`, because `StopAsyncIterator` exception is used in Python language itself to notify the termination of an iteration.

This commit fixes it by making the hook ignores `StopAsyncIterator` when `max_failure` is enabled.


The following code now works.

```
from spdl.pipeline import PipelineBuilder


def generator_fn(arg):
    for i in range(arg):
        yield i


def main():
    src = list(range(100))
    pipeline = (
        PipelineBuilder()
        .add_source(src)
        .pipe(generator_fn)
        .add_sink()
        .build(num_threads=1, max_failures=1)
    )
    with pipeline.auto_stop():
        for i in pipeline:
            print(i, flush=True)


if __name__ == "__main__":
    main()
```

Differential Revision: D80941146