Fix the failure monitoring in ordered input
In OrderedPipe, to preserve the order, an internal queue is created.
Then two coroutines in a form of producer/consumer are created.
The producer creates pipe task on input values, and put them in the
internal queue.
The consumer receives a task from the queue, check its status and
propagate the result to the output queue.

The previous implementation attached the failure counter to each task,
and the producer was reacting to too-many-failures, then
the consumer was blindly processing the task.

This allowed a case where one task to hit the maximum allowed number
of failures, but subsequent tasks (already put in the internal queue) to
be processed, breaking the promise of ordered output.

To address this, this commit changes the place where the monitoring
happens. Now FailCounter is used at the point where the consumer
checks the result of tasks. By doing so consumer is now aware of
failure counting, and it can react to the situation where the maximum
failure is hit, and discard any subsequent task even they are successful.