Fix FailCounter's global failure count mechanism (#950)
In
https://github.com/facebookresearch/spdl/commit/f8b9904b84e2cecc9a846e5544fbaebcfb0bd2dd,
the `_FailCounter.increment` method was changed from class method to
instance method.

It turned out that `+=` operator in instance method creates an instance
attribute.
So the global failure counter value was not shared among the pipe
stages.

I thought such pattern is caught by the tests but it seems not.
This commit fixes it and add a test to verify that.