Add explicit dependency structure to Pipeline coroutines (#959)
This commit introduces _Node structure which is used internally to
represent the relationship of coroutines.

Currently, Pipeline only supports a chain. When we add support for
merging multiple Pipeline, we need a proper way to handle cancellation
for upstream tasks which might fork.

The _Node class helps implementing this.

The newly added unit test shows that it can handle complex cases (such
as nested merges, and 3-way merges).