Convert PipelineConfig to graph structure then build
Currently, the graph representation is constructed as components
of PipelineConfig are converted into coroutines.

This works fine for simple straight chain structure, but when
we start dealing with Y-shaped graph structure, with potential
nesting, the recursive function calls becomes complicated and
the function signature for build function becomes too verbose.

It is simpler to converting the PipelineConfig into a graph structure
first because graph retains the necessary dependency information,
so recursive operation is carried out in graph traversal, but not
in the coroutine construction.

This makes it easier to add merge mechanism.