Add self-diagnostic mode to pipeline (#982)
This commit adds self-diagnostic mode, as outlined in
https://github.com/facebookresearch/spdl/issues/903

When the environment variable `SPDL_PIPELINE_DIAGNOSTIC_MODE` is set,
the build_pipeline function builds the pipeline in self-diagnostic mode.

In this mode, the pipeline benchmark the stage function one-by-one, and
reports the performance.

This is useful when one needs to profile the pipeline in prod
environment without changing the code.

We will add a mechanism to configure callback and hook in
self-diagnostic mode.