Increase test time out
On GitHub, `test_iterate_in_subprocess_buffer_size_64` fails
sporadically, on Windows and Linux (both aarch64 and x86_64).

```
>   assert int(f.read()) == 9
           ^^^^^^^^^^^^^
E   ValueError: invalid literal for int() with base 10: ''
```

As far as I recall, the value is always empty when test fails,
so perhaps it's timeout.

This commit increases the timeout to 10 seconds.