Add interface file
We want to properly support static type checking.

But we have constraints
1. `import spdl.io` must succeed without having FFmpeg and/or
   CUDA runtime installed
2. Classes defined in extension module, such as `AudioCodec`
   must be a concrete class under `spdl.io`, so that one can
   use `isinstance`.

Currently, we have `spdl.io._type_stub.py`, which is a fake
classes used to satisfy the static checker, while supporting
only point 1.

To support point 2, we need to separate the behavior for
type-checking and actual program execution.

This commit is the first step.
It introduces the top-level interface file so that static
type-checker takes advantage of the type stubs, which allows
to load symbols without loading extension modules.