Remove type stub
This is a continuation of https://github.com/facebookresearch/spdl/pull/1131

In this commit, the mock stub file is removed, and the docstrings
are moved th teh actual classes bound in C++.

At runtime, the attributes like `AudioCodec` is not retrieved
until they are requested by caller.

When an attributed is requested but SPDL failed to laod the
target due to missing dependencies, then a mock object is returned.

If the call site calls the mock object, it will show the failure.