Type update [2/3] Import extension classes only when type checking (#1132)
We want to properly support static type checking.

But we have constraints
1. `import spdl.io` must succeed without having FFmpeg and/or CUDA
runtime installed
2. Static type-checkers need to work without having FFmpeg and/or CUDA
runtime installed.
3. Classes defined in extension module, such as `AudioCodec` must be a
concrete class under `spdl.io`, so that one can use `isinstance`.

Currently, we have `spdl.io._type_stub.py`, which is a fake classes used
to satisfy the static checker,
this supports point 1 and 2 but not 3..

To support point 3, we need to separate the behavior for type-checking
and actual program execution.

- https://github.com/facebookresearch/spdl/pull/1131
This commit introduces the top-level interface file so that static
type-checker takes advantage of the type stubs,
which allows to load symbols without loading extension modules.
- https://github.com/facebookresearch/spdl/pull/1132
In this commit, we use `TYPE_CHECKING` so that relevant classes are
imported only when type-checking.
We apply postponed evaluation of annotations (stringify the annotation)
  so that the classes are only accessed during the type-checking.
(and type checkers actually access the stub file, and not the extension
modules)
- https://github.com/facebookresearch/spdl/pull/1133
In this commit, the mock stub file is removed, and the docstrings are
moved to the actual classes bound in C++.
At runtime, the attributes like `AudioCodec` is not retrieved until they
are requested by caller.
When an attributed is requested but SPDL failed to load the target due
to missing dependencies,
then a placeholder object is returned. If the call site calls the mock
object, it will show the failure.

This way, the extension module is not required at the time of runtime
checking.
Importing attributes from `spdl.io` will return something so if
`spdl.io` is
imported by third-party library but the user does not require it (thus
not having
the proper dependency), the user code works.
And if one wants to use `spdl.io`, given dependencies are installed,
types are
properly propagated to user code so that they can use `isinstance`.
Finally, the documentation has proper docstrings.