Use the stream time base when seeking (#1152)
Summary:
We found a case where seeking does not work well, and this can be
mitigated if we use the stream time base instead of global time base.

Unfortunately, the issue seems to be data-dependent and I was not able
to create a test data that reproduces the issue.

Meta folks can check D87452802 for repro.

Differential Revision: D87556043