Add build callback (#1162)
This commit adds a callback function to `build_pipeline`.

The purpose of the callback is so that we can log the detail of pipeline
without changing the way SPDL is integrated.

This allows to inject logging mechanism across code base without
changing the individual projects that use SPDL.

This functionality is intended for adding logging to the code that does
not have access to the part that calls `build_pipeline` function. So we
do not add equivalent `callback` argument to `build_pipeline` function.
(If you are calling `build_pipeline`, you can achieve the same effect by
calling your function directly by providing PipelineConfig object.)