Fix the integer overflow when converting PTS to rational type
AVRational stores the data as a tuple of int. (int32_t)
The packet PTS is stored as int64_t.

There are cases where the `AVPacket::pts` value is larger than
`INT32_MAX`, and in such case we cannot simply cast the variable
type. We need to perform reduction, otherwise it causes interger
overflow.

Test file is also added but there is no OSS setup yet.