Fix the integer overflow when converting PTS to rational type (#1185)
`AVRational` stores the data as a tuple of int (`int32_t`). 
`AVPacket::pts` stores PTS as `int64_t`.

There are cases where the `AVPacket::pts` value is larger than`INT32_MAX`.
In such case, we cannot simply cast the variable type.
We need to perform reduction, otherwise it causes interger overflow.

The test will be added in a subsequent PR.