Update the way timestamp is handled (#1153)
1. When comparing the timestamp and PTS, convert them into AVRational
instead of double, so that the comparison is more accurate.
2. Use the same timestamp filtering mechanism in Packets and decoding.
3. Do not use the `trim` filter for timestamp mechanism. Instead filter
manually, using the above mechanism.

The new implementation gives a result closer to Python's slice notation.
For example in one of the test, we used `timestamp=(2.88, 2.89)`. The
`trim` filter will discard the frame at 2.88, while the new
implementation keeps it.