Ensure upstreams are cleaned up when Merge stage finishes (#1207)
Introduces a mechanism to ensure that the stages upstream to merge
operations are
cleaned up when a merge operation is completed.

This enables custom merge operations to exit early based on their own
logic
while ensuring proper resource cleanup.

**Implementation detail**

We introduce a `_StageCompleted` exception type, which is sort of
analogous to Python's
built-in `StopIteration` exception, and the merge stage to throw it at
the end.
This signals the Pipeline execution mechanism to clean up the upstream
stages.
Then we filter out the instances of `_StageCompleted` from the list of
errors that
should be propagated to the front end, so as not to treat it as a real
error.

Note: In a subsequent PR, we get rid of `_StageCompleted` and simplify
the logic.