Do not specify cmake and ninja in pyproject.toml (#1211)
For a while we had `cmake` and `ninja` specified in `pyproject.toml` of
spdl-io package.

Sometimes with `uv`, the build process cannot find `ninja` for a reason
I don't know, and the issue parsists.

It's pretty annoying when local dev, so we do not specify these within
build, and rely on the host to have these installed somewhere.