Handle RASL frames on best-effort basis (#1236)
RASL (Random Access Skipped Leading) frames in HEVC (H.265) refers to
special pictures that appear after a key random access point (like a CRA
picture) but depend on pictures that come before the random access point
in coding order, making them undecodable if you start from the CRA
picture; they are typically skipped during playback to enable clean,
immediate starts at CRA points, unlike RADL (Random Access Decodable
Leading) pictures which are decodable.

This was the case for `spdl.io.demux_video` as well.

This commit changes the behavior slightly by performing additional seek
when RASL is suspected.
The additional seek is performed with the timestamp value of ``DTS of
the packet - 1``, so that it will seek to the keyframe before the
keyframe the initial seek pointed.