[libspdl] Remove the original NV12 color conversion function
With the changes introduced in
https://github.com/facebookresearch/spdl/pull/1271 ,
NVDEC decoder only returns batched frames, and corresponding
NV12 conversion function was added in
https://github.com/facebookresearch/spdl/pull/1269 .

These changes make the original NV12 color conversion function
(the one takes a list of single frames) obsolete and disconnected
from the rest of the APIs.

So this commit removes the funciton.